requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
elevenlabs==1.50.2
eval_type_backport==0.2.0
griffe==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
logfire-api==2.7.1
moviepy==2.1.2
numpy==2.2.3
openai==1.68.0
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
python-dotenv==1.0.1
replicate==1.0.4
requests==2.32.3
sniffio==1.3.1
sounddevice==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
